+++ Tests/PlayerTests/PlayerDeallocationTests.swift
@@ -1,0 +1,25 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+@testable import Player
+
+import AVFoundation
+import Nimble
+import XCTest
+
+@MainActor
+final class PlayerDeallocationTests: XCTestCase {
+    func testPlayerDeallocation() {
+        let item = AVPlayerItem(url: TestStreams.onDemandUrl)
+        var player: Player? = Player(item: item)
+
+        weak var weakPlayer = player
+        autoreleasepool {
+            player = nil
+        }
+        expect(weakPlayer).to(beNil())
+    }
+}
