+++ Tests/PlayerTests/PlayerItemPublishersTests.swift
@@ -8,7 +8,6 @@
 
 import AVFoundation
 import Circumspect
-import Combine
 import Nimble
 import XCTest
